float zu long

+++ src/dataloader.py
@@ -33,7 +33,7 @@ class Kaggle50K(Dataset):
         image_path = self.image_paths[idx]
         label_str = self.labels[idx]
         label = self.labels_to_idx[label_str]
-        label = torch.FloatTensor([label])
+        label = torch.LongTensor([label])
 
         image = Image.open(image_path).convert("RGB")
         if self.transform:
